Implements a python-based microservice
This commit implements a python-based microservice with the help of
FastAPI, Pydantic and Uvicorn.

It implements one endpoint, which checks if a target value is present in
two given lists.

Additionally, it containerizes the above mentioned web application in a
docker container to make it agnostic to its deployment environment and
enable deployment on all docker container compatible environments.

To make sure that the code is in good shape and also works like
expected, tests are provided along with tools like GitHub Actions,
Linting and Formatting Checker and also static type checking for python.